html/rendering/bindings/the-marquee-element/marquee_loop.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: marquee_loop</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/obsolete.html#the-marquee-element-0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div>
      <marquee id="test1" loop="a">Test1</marquee>
      <marquee id="test2" loop="-2">Test2</marquee>
      <marquee id="test3" loop="2">Test3</marquee>
    </div>
    <script>
      setup({explicit_done:true});

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test1");
        assert_equals(mq.loop, -1, "The value of loop should be -1.");
      }, "marquee_loop_string");}, 100);

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test2");
        assert_equals(mq.loop, -1, "The value of loop should be -1.");
      }, "marquee_loop_less_than_1");}, 100);

      setTimeout(function() {test(function() {
        var mq = document.getElementById("test3");
        assert_equals(mq.loop, 2, "The value of loop should be 10.");
      }, "marquee_loop_normal");}, 100);

      setTimeout(function(){done()}, 120);
    </script>
  </body>
</html>
